+++ Tests/Unit/Domain/Repository/Fixtures/ReadOnlyRepository.php
@@ -2,7 +2,7 @@
 
 namespace OliverKlee\Oelib\Tests\Unit\Domain\Repository\Fixtures;
 
-use OliverKlee\Oelib\Domain\Repository\Traits\ReadOnlyTrait;
+use OliverKlee\Oelib\Domain\Repository\Traits\ReadOnly;
 use TYPO3\CMS\Extbase\Persistence\Repository;
 
 /**
@@ -12,5 +12,5 @@ use TYPO3\CMS\Extbase\Persistence\Repository;
  */
 class ReadOnlyRepository extends Repository
 {
-    use ReadOnlyTrait;
+    use ReadOnly;
 }